Add meta descriptions on blog post pages
I really dislike this duplication, but I couldn’t quickly figure out
how to get it to populate that from a data file. I will likely come
back to this later. So much of this project is configured *just so*
that I hate to cut corners anywhere. I absolutely love this new site.