Maybe this will fix sorting?
If I understand, creating the collection in `.eleventy.js` should make
it unsorted. I don’t know why I needed to reverse the collection for
the nav for it to show up in the right order, but if it works, I’ll be
happy. And regardless, thanks to Mathieu Huot in Discord.

https://www.11ty.dev/docs/collections/